Refactored the help system structure
